updating test imports and middleware test
