+++ tests/cp_mock.py
@@ -5,9 +5,7 @@ import sys
 sys.path.append("emulator/cp/")
 sys.path.append("flight/")
 
-# Mock other modules
-sys.modules["micropython"] = __import__("micropython_mock")
-sys.modules["ulab"] = __import__("ulab_mock")
-
 # Alias 'hal' to 'emulator'
 sys.modules["hal"] = importlib.import_module("emulator")
+
+import hal.cp_mock
